==> notebook/1_generate_dummy_data.ipynb <==
#### Notebook to generate the dummy data and saving to CSV

import pandas as pd
import numpy as np

##### First generate the 10 dummy features

dummy_data = {
    'soil_color': np.random.choice(['dark brown', 'reddish', 'light brown'], 1000),
    'soil_ph': np.random.uniform(5.5, 7.5, 1000),
    'soil_n': np.random.uniform(10, 50, 1000),
    'soil_p': np.random.uniform(10, 50, 1000),
    'temp': np.random.uniform(25, 35, 1000),
    'rainfall': np.random.uniform(0, 300, 1000),
    'forecast_temp': np.random.uniform(25, 35, 1000),
    'forecast_rainfall': np.random.uniform(0, 300, 1000),
    'crop_type': np.random.choice(['wheat', 'corn', 'rice'], 1000),
    'plant_health': np.random.choice(['healthy', 'yellowing', 'wilting'], 1000)
}

df = pd.DataFrame(dummy_data)
df.head()

##### Then generate the target variable


def calculate_optimal_fertilizer(row):

    # Base amount of fertilizer in kg/ha
    base_amount = 100

    # Adjust fertilizer amount based on soil nitrogen levels
    nitrogen_adjustment = (50 - row['soil_n']) * 0.5

    # Adjust fertilizer amount based on soil phosphorus levels
    phosphorus_adjustment = (50 - row['soil_p']) * 0.3

    # Add more fertilizier if ph level is not good  
    ph_adjustment = 10 if row['soil_ph'] < 6 or row['soil_ph'] > 7 else 0

    # Adjust the fertilizer amount based on weather conditions
    weather_multiplier = 0.9 if 25 <= row['temp'] <= 30 and 50 <= row['rainfall'] <= 150 else 1

    # Adjust the fertilizer to use based on crop type 
    crop_multipliers = {'wheat': 1.0, 'corn': 1.2, 'rice': 0.9}

    # Calculate the final fertilizer amount to use 
    final_fertilizer_amount = (base_amount 
                               + nitrogen_adjustment 
                               + phosphorus_adjustment 
                               + ph_adjustment) * weather_multiplier * crop_multipliers[row['crop_type']]
    
    return final_fertilizer_amount

df['optimal_fertilizer_amount'] = df.apply(calculate_optimal_fertilizer, axis=1)

df = df.round(2) 

df.head()

#### Introduce missing values and outliers in the dummy data

# Introduce missing values for temperature and rainfall
missing_percentage = 0.003 

for col in ['temp', 'rainfall']:

    num_missing = int(len(df) * missing_percentage)
    missing_indices = np.random.choice(df.index, size=num_missing, replace=False)
    df.loc[missing_indices, col] = np.nan

# Introduce outliers for soil nitrogen and soil phosphorus 
outlier_percentage = 0.001  

for col in ['soil_n', 'soil_p']:

    num_outliers = int(len(df) * outlier_percentage)
    
    outlier_indices = np.random.choice(df.index, size=num_outliers, replace=False)
    
    # Outliers are set at 100 times of the value
    df.loc[outlier_indices, col] = df[col].max() * 100  

# Check if the missing values / outliers were introduced as intended
df.describe()

#### Save dummy data to csv

df.to_csv('../data/dummy_data.csv', index=False)
