==> notebook/3_model_findings_and_recommendations.ipynb <==
import joblib
import pandas as pd

#### Load the trained model

model = joblib.load("..\model\model_pipeline.joblib")

#### Give recommendations based on example scenarios

# Scenario 1: Yellowing Wheat with Low Rainfall and Low Soil Nutrient Levels
condition_1 = pd.DataFrame([{
    'soil_color': 'dark brown',
    'soil_ph': 5.8,
    'soil_n': 15.0,
    'soil_p': 20.0,
    'temp': 25.0,
    'rainfall': 10.0,
    'forecast_temp': 27.0,
    'forecast_rainfall': 60.0,
    'crop_type': 'wheat',
    'plant_health': 'yellowing'
}])

optimal_amount_1 = model.predict(condition_1)
print(f"Optimal Fertilizer Amount for Scenario 1: {optimal_amount_1[0]} kg/ha")

# Scenario 2: Healthy Rice with High Rainfall and High Soil Nutrient Levels
condition_2 = pd.DataFrame([{
    'soil_color': 'reddish',
    'soil_ph': 7.0,
    'soil_n': 45.0,
    'soil_p': 50.0,
    'temp': 30.0,
    'rainfall': 200.0,
    'forecast_temp': 32.0,
    'forecast_rainfall': 210.0,
    'crop_type': 'rice',
    'plant_health': 'healthy'
}])

optimal_amount_2 = model.predict(condition_2)
print(f"Optimal Fertilizer Amount for Scenario 2: {optimal_amount_2[0]} kg/ha")

# Scenario 3: Wilting Wheat with Very Low Nutrient Levels and Low Rainfall
condition_3 = pd.DataFrame([{
    'soil_color': 'light brown',
    'soil_ph': 5.5,
    'soil_n': 10.0,
    'soil_p': 10.0,
    'temp': 25.0,
    'rainfall': 10.0,
    'forecast_temp': 26.0,
    'forecast_rainfall': 15.0,
    'crop_type': 'wheat',
    'plant_health': 'wilting'
}])

optimal_amount_3 = model.predict(condition_3)
print(f"Optimal Fertilizer Amount for Scenario 1: {optimal_amount_3[0]} kg/ha")
